Delete integrated tests + update other ones
